Refactor: Remove interfaces from NewContact and change import locations in apiCalls and move interfaces to Interfaces.tsx.
