+++ src/Interfaces.tsx
@@ -48,57 +48,53 @@ export interface JobApplication {
   updated_at: string;
 }
 
-
-
-
-
   export interface DashBoardDattaI{
-          job_applications: [ null |
-              {
-                  id: number;
-                  position_title: string;
-                  date_applied: string;
-                  status: number;
-                  notes: string;
-                  job_description: string;
-                  application_url: string;
-                  created_at: string;
-                  updated_at: string;
-                  company_id: number;
-                  user_id: number;
-              }
-          ],
-          contacts: [ null |
+    job_applications: [ null |
+        {
+            id: number;
+            position_title: string;
+            date_applied: string;
+            status: number;
+            notes: string;
+            job_description: string;
+            application_url: string;
+            created_at: string;
+            updated_at: string;
+            company_id: number;
+            user_id: number;
+        }
+    ],
+    contacts: [ null |
 
-              {
-                  id: number;
-                  first_name: string;
-                  last_name: string;
-                  email: string;
-                  phone_number: string;
-                  notes: string;
-                  created_at: string;
-                  updated_at: string;
-                  user_id: number;
-                  company_id: number;
-              }
-          ],
-          companies: [ null |
+        {
+            id: number;
+            first_name: string;
+            last_name: string;
+            email: string;
+            phone_number: string;
+            notes: string;
+            created_at: string;
+            updated_at: string;
+            user_id: number;
+            company_id: number;
+        }
+    ],
+    companies: [ null |
 
-              {
-                  id: number;
-                  user_id: number;
-                  name: string;
-                  website: string;
-                  street_address: string;
-                  city: string;
-                  state: string;
-                  zip_code: string;
-                  notes: string;
-                  created_at: string;
-                  updated_at: string ;
-              }
-          ]
+        {
+            id: number;
+            user_id: number;
+            name: string;
+            website: string;
+            street_address: string;
+            city: string;
+            state: string;
+            zip_code: string;
+            notes: string;
+            created_at: string;
+            updated_at: string ;
+        }
+    ]
 
   }
 
@@ -135,4 +131,21 @@ export interface JobApplication {
         notes: string;
       };
     };
+  }
+
+  export interface FormInputData {
+    firstName: string;
+    lastName: string;
+    email: string;
+    phoneNumber: string;
+    companyId?: number | null;
+    notes: string;
+  };
+
+  export interface NewContact {
+    first_name: string,
+    last_name: string,
+    email: string,
+    phone_number: string,
+    notes: string,
   }
