+++ src/apiCalls.tsx
@@ -1,5 +1,4 @@
-import { FormInputData } from "./components/contacts/NewContact"
-import { NewContact } from "./components/contacts/NewContact"
+import { FormInputData, NewContact } from "./Interfaces"
 const apiURL = process.env.REACT_APP_BACKEND_API_URL
 const backendURL = `${apiURL}api/v1/`
 /*-----------------------------------// GET //--------------------------------------*/
